QmlMessageTrace: detect and combine repeated sections of backtraces
If there are multiple tracepoints, it happens quite often that the
most recent backtrace contains the backtrace that we already captured
in handling the previous tracepoint.  If the time difference is small
enough and we find such a duplicated sequence of methods, don't repeat
the Message output in the trace QML: pick up where the previous trace
left off.

Change-Id: I32f47430ab44fd7570569e0f347784d23197a3bc
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>