install QML for viewing Sequence diagrams; remove uq script
Now you can simply view the diagram QML with the qml runtime.
The module is now renamed to Qt.labs.UmlQuick, with the Sequence
(MessageTrace) diagram components being in Qt.labs.UmlQuick.Sequence
and the trace instrumentation being in Qt.labs.UmlQuick.Trace.
The diagram in UML terminology is called simply a Sequence diagram,
whereas the act of generating one is to trace messages.

Change-Id: I4d7db9eecc32af8b57c2c4242bb7d4568b777e28
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>