Connect to QCoreApp::aboutToQuit rather than writing in the dtor
We should not rely on destruction order. When the application is
quitting, it may be that by the time the QmlMessageTrace dtor is called,
other objects are already destroyed, and thus introspecting object
pointers stored in struct Message instances will be dangerous (for
example calling o->metaObject()->className() crashed in one case).
The aboutToQuit signal is emitted a bit earlier.

Change-Id: I4e6bf73586f39ff24f825fb9ce2c8f56d4a02ee7
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>