Convert to cmake
There are two modules: one to generate traces at runtime, and one
containing components used in the generated QML, needed to view the
traces. We use add_subdirectory; there is some redundancy between the
CMakeLists.txt in the subdirectories, but they are difficult to combine
because of the use of per-module variables like module_dir, module_uri
and module_version.

We use PLUGIN_TARGET to get the backing code linked into the plugins (as
was normal in Qt 5): Qt.labs.UmlQuick.Trace is a development-oriented
module, generally used with temporarily-instrumented QML files, not
meant for application deployment. Generated QML traces are generally
viewed with the qml tool, so again it doesn't seem beneficial to have
Qt.labs.UmlQuick.Sequence available in a separate backing library.

Task-number: QTBUG-111946
Change-Id: Ia0f0b3bd43f64d0e0cfc67dd1e93165626a8c28b
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>